Force help to used a fixed column width for the names
The help formatting for options, arguments and sub-commands splits the
output into two columns, the first for the name and the second for the
description. If the name exceeds the maximum column width then it will
be placed onto its own line and the description will be on the
following line, otherwise the first line of the description will be on
the same line as the name.

The actual column width is variable and depends on the maximum length
of name in each section, capped to the maximum column width. The
problem with that is until the actual column width is the same as the
maximum column width, adding a new name that is longer than all the
previous names will cause the formatting of those names to change
causing unnecessary churn, making changes harded to review and more
likely to conflict.

This change effectively forces the help to have a fixed column width
for all options, arguments and sub-commands.

Bug: 286023667
Test: ./gradlew --rerun-tasks
Change-Id: Iceaa1cf0253ff9cc3859b52ca9c0ef57d178b2eb