Add android-jars-to-signatures sub-command
Previously, the updating of the existing android signature files (in
`prebuilts/sdk/<api>/public/android.txt`) was performed by running:
```
    metalava --write-android-jar-signatures *<android source dir>*
```

This change replaces that with a new sub-command:
```
    metalava android-jars-to-signatures *<android source dir>*
```

It replaces the legacy option (rather than having both overlapping)
because it is not used on a regular basis. It is only run, as and when
needed.

Bug: 286023667
Test: ./gradlew --rerun-tasks
      metalava android-jars-to-signatures $ANDROID_BUILD_TOP
Change-Id: Ib39aa29fe82793a6bc5ec130096dc787f4ccf751