Ensure global options is initialized correctly
Change https://r.android.com/2633050 moved the parsing of the command
line arguments array from the initializer to the `parse()` method. As
a result the global default `options` no longer parsed an empty array
of arguments. The intent was that an `Options()` object after the
change would be identical to an `Options(emptyArray())` before the
change. While no tests were broken it appears as though that was not
true for a couple of properties as previously the initializer would:
* set `showUnannotated` to true if `showAnnotations` was empty.
* set `baselineBuilder.file` to the default baseline file if it
  existed.

This change fixes that by calling `parse(emptyArray())`. It is needed
to ensure that follow up changes which rely on the global `options`
work correctly and do not have to wait for the options to be cleaned
which is in progress but will take a while.

Bug: 286023667
Test: ./gradlew --rerun-tasks
Change-Id: I7de08fcc6b77002aa75243ca9fe38330095104b2