Add ability to run local metalava against a pinned version of androidx
Adding an integration project so we can easily benchmark metalava
invocation performance for a call that is similar to androidx usage
of metalava. The project is not included in settings.gradle because
stable version of Intellij that is used to develop metalava is unable
to work with the latest Android Gradle Plugin that we use to do this
integration.

Test: INTEGRATION=true ./gradlew integration:run --rerun
      seeing ~22s execution
Change-Id: Ib302130ce95005aff305a1822d4a5f45044170c8