Push Item.psi(): PsiElement? down to PsiItem.psi(): PsiElement
The `Item.psi(): PsiElement?` method made the abstract model dependent
on IntelliJ classes which is a massive and unnecessary dependency.

The first part of the change was to add overloads of `psi()` on
`Psi*Item` classes to use the class specific subclass of `PsiElement`.
e.g. `PsiClassItem.psi(): PsiClass`. That reduces the amount of
additional casting needed after calling `psi()`.

Then, calls to `Item.psi()` were fixed using whichever of the following
rules applied first:
1. The receiver type was switched to a `Psi*Item`, e.g. if
   `containingClass(): ClassItem` can be overridden by
   `containingClass(): PsiClassItem` then `containingClass().psi()`
   does not need changing.
2. An unconditional down cast was done when it was clear that the
   receiver had to be of the target type.
3. A conditional down cast was done otherwise.

The target type for a down cast was the most specific type possible,
e.g. when the receiver type was `ClassItem` then it would be cast to a
`PsiClassItem`.

Bug: 288398409
Test: ./gradlew
Change-Id: Id02e1a976db6afca1728073722725ef3c3b6b59c