Make PsiConstructorItem.psi() consistent with PsiMethodItem
Previously, the `PsiConstructorItem.psi()` method would sometimes
return a `PsiMethod` and sometimes returns a `PsiClass` depending on
whether the underlying constructor was explicitly defined in the source
code, or was an implicit/synthetic constructor. The reason for that was
because the `PsiElement` returned by `psi()` was used to determine the
location but the `PsiElement` for an implicit/synthetic constructor did
not provide the information needed to do that.

Recent changes added an explicit `location()` method that is used
instead of the `psi()` method and so this change overrides that method
to provide the location instead. That allows the `psi()` method to
be removed and fallback to the `PsiElement` returned by the parent.
That did necessitate an extra check in `DocAnalyzer` but it does mean
that code that accesses the `psi()` can rely on a `PsiMethodItem`
(and the `PsiConstructorItem` sub-class) returning a `PsiMethod`
object.

Bug: 288398409
Test: ./gradlew
Change-Id: I54ee32edc466e878f5bc56b55a7b5cbf994c387b