Add Location abstraction
The `Reporter` classes provides methods to report issues related to
`File`, `PsiElement` and `Item` objects. Unfortunately, that involves
a lot of duplication and inconsistencies. The use of `PsiElement`
makes it impossible to use `Reporter` (which is a generally useful
class) without dragging in the whole Psi dependencies.

The `Baseline` is similar as it needs to provide baseline keys for
the issues reported by `Reporter` so it too has dependencies on Psi
classes.

This change adds `Location` and `BaselineKey` abstractions that will
insulate the `Reporter` and `Baseline` classes from directly depending
on the Psi classes (among others) and allow for the inconsistencies and
duplication to be addressed.

This adds support for creating `Location` and `BaselineKey`
abstractions for a `File`. Following changes will cover `PsiElement`
and `Item`.

Bug: 288398409
Test: ./gradlew
Change-Id: I5544b35eae61fb7950990230ac10045b48361509