Report File issues the same way as PsiElement issues
Previously, when reporting issues with a `File` the path would not be
made relative to `rootFolder` (the test temporary directory root). As a
result tests for such issues relied on the `DriverTest.cleanupString()`
method to remove the temporary directory from the test output. The
consequence of that was those tests would end up with `TESTROOT` in the
output.

That is different to `PsiElement` which relatives the path in the
report against `rootFolder`.

This change applies the same transformation to the `File` path that the
`PsiElement` reporting code applies to its path. That required some
tests to be updated.

Bug: 288398409
Test: ./gradlew
Change-Id: I2184731a93548e4f5f303e6b8e014992f173caa9