Reset options before each BaseCommandTest
Some commands being tested rely on the global `options` variable being
set to its default value. This change ensures that the behavior of
those commands is not affected even if it runs after a test that
modifies the global `options`. That ensures that the commands behave
the same during testing as they do when run as a standalone process.

Bug: 286023667
Test: ./gradlew --rerun-tasks
Change-Id: I7aa539b9c3dd0b4f487a6727288bea7d413a797a