Remove ClassItem.publicSuperClass()
Previously, the `ClassItem.publicSuperClass()` was only called from
`MethodItem.filteredThrowsTypes()`. The `publicSuperClass()` filtered
the super class hierarchy to only include `ClassItem`s that were needed
for the `options.docLevel`. However, the `filteredThrowsTypes()` then
filtered the `ClassItem` using the supplied `predicate` which is always
`filterReference`. That predicate is used everywhere to determine which
classes can be referenced from an API and so is much more accurate than
the simple `checkLevel()` used in `publicSuperClass()`.

As long as `checkLevel()` is less restrictive than `filterReference`
then it would be safe to remove the call to `checkLevel()`. While it is
theoretically possible that the `checkLevel()` would filter out a class
that would not be filtered out by `filterReference` all the testing
done shows that does not seem to be happening in practice.

Therefore, this change replaces any calls to `publicSuperClass()` with
calls to `superClass()` and removes the `publicSuperClass()` method.

Bug: 289919515
Test: ./gradlew
Change-Id: Ia3c547ec5e1447d2afa9a98ad1292574e92329ec