Remove inconsistent TextCodebase override of compareWith()
Previously, the `TextCodebase` class overrode `compareWith()` and
switched the order of the `other` and `this` parameters when it called
`CodebaseComparator().compare(...)`. In the `Codebase.compareWith()`
method the `this` parameter was passed as `new` and the `other`
parameter was passed as `old`. However, in `TextCodebase.compareWith()`
they were reverted so the `this` parameter was passed as `old` and the
`other` parameter was passed as `new`. It makes no sense of the order
to defend on the type of the `Codebase` on which the `compareWith()`
method is called.

Given that the `Codebase.compareWith()` method is only called from the
`Driver.migrateNulls()` method and it calls `previous.compareWith()` it
seems as though the order used by `TextCodebase` is the correct order.
That is confirmed by testing. Tests fail if
`TextCodebase.compareWith()` is removed but they pass if its method
body is pulled up to `Codebase.compareWith()`.

So, this change does exactly that.

Bug: 288550547
Test: ./gradlew
      # Will also test on Android and AndroidX builds
Change-Id: I5af4ed7f110d863d5afa4a0dffe7b3809f2a1c94