Move ItemVisitor and TypeVisitor into model package
The `visitors` package contains other classes which have dependencies
that cannot be in the `model` package. There is no reason for the
`ItemVisitor` and `TypeVisitor` to be in a separate package to the
items so moving them into the `model` package is the cleanest way
to avoid the `model` package depending on other classes in the
`visitors` package.

Bug: 290775876
Test: ./gradlew
Change-Id: Ic6d27cee28156888787957acc39b5c8b86a75197