Add NoOpAnnotationManager for use in ApiFile
The `ApiFile` has a couple of `parseApi()` methods that are used by
external code and text model specific tests that need an
`AnnotationManager`. Previously, they used the
`DefaultAnnotationManager` but that really needs to be moved out of the
model as it has a lot of logic that is specific to the users of the
model and is not intrinsic to the model. This change adds a new
`NoOpAnnotationManager` that provides limited but sufficient support
for use by those methods.

A follow up change will move `DefaultAnnotationManager` out of the
`model` package.

Bug: 288550547
Test: ./gradlew
Change-Id: I30f2e5105567c1a086de78de2234e72550abbd17