Support varargs in TypeItem.toSlashFormat()
The type string of a varargs parameter ends with `...` but needs to be
treated as an array at runtime. Previously, the `toSlashFormat()` did
not handle it and so it treated types like `int...` as if they were an
object and treated types like `Object...` as if they were an inner
class with empty string as the name.

This change adds a test for those cases and fixes the issue. The test
is structured so that it is run against both a text model and a psi
model as the behavior needs to be consistent between the two.

Bug: 287241408
Test: ./gradlew
Change-Id: I5284927f6bd6f070d402656c246d2ebea0a0c2ff