Make Psi/UAnnotationValue sub-classes extend DefaultAnnotationValue
Previously, the `Psi/UAnnotationValue` sub-classes had to duplicate a
of functionality from the `DefaultAnnotationValue` sub-classes. This
change modifies them to extend a `DefaultAnnotationValue` sub-class
which allows the removal of that duplicated code.

A follow up change will remove the classes which are no longer needed
and replace them with the appropriate `DefaultAnnotationValue`
sub-class.

Bug: 291562733
Test: ./gradlew
Change-Id: Idc46ff6d939563d2b14f41df491c0a9623337fcb