Add TextBackedAnnotationItem.create factory method
Previously, the `TextBackedAnnotationItem` constructor took a `source`
parameter which it then parsed into `originalName` and `attributes`.
That made it harder to move some of the duplicate functionality of that
class up into `DefaultAnnotationItem`.

This change moves the parsing code out of the constructor and moves it
into a new `TextBackedAnnotationItem.create` factory method. The
constructor then just takes the `originalName` and `attributes` as
parameters.

Bug: 291562733
Test: ./gradlew
Change-Id: I366ca4de1f00ba4bee0be9d35f6ec9c150c9c558