Add DefaultAnnotationItem.toSource()
Previously, `XmlBackedAnnotationItem` and `TextBackedAnnotationItem`
both provided their own implementations of the `toSource()` method.
This change merges those implementations into `DefaultAnnotationItem`.

The `XmlBackedAnnotationItem` implementation constructed the source
from the different parts so the `DefaultAnnotationItem` implementation
does too.

The `TextBackedAnnotationItem` essentially preserved its input which
did not include the default `value` attribute if there was only one
attribute specified. So, this handles that too.

A minor change was needed to a test that had some additional spacing
which this change did not preserve. That made the test slightly more
realistic as signature files do not contain that spacing.

Bug: 291562733
Test: ./gradlew
Change-Id: Idb6b16b4cbd06f4c5f3ee5b3cbca152ed72562a2