Stop XmlBackedAnnotationItem mapping annotation name twice
Change https://r.android.com/1817102 switched from `qualifiedName()`
function to `qualifiedName` property and in the process it accidentally
caused the `toSource()` method to map the name of an annotation that
was already mapped.

Prior to the change `qualifiedName` was a private property that was
actually what is now called `originalName`. As part of that change
the `toSource()` method should have been updated by either:
1. Removing the `qualifiedName` local altogether and using the
   `qualifiedName` property. This would have been the best.
2. Replacing the use of the `qualifiedName` property with
   `originalName`.

This change applies #1.

Bug: 291562733
Test: ./gradlew
Change-Id: I824628b20258d65573c95659acb9b3a79d22f2de