Stop switching cloned method to public
The `PsiClassItem.createMethod(template: MethodItem)` creates a new
method that is supposed to be a clone of the original method, in a
sub-class of the containing class of the template method. If the
template method has unresolved type variables, then if possible they
are resolved by the type variable bindings of the sub-class.

e.g. If the parent class was:
```
class Parent<T> {
   public void inherited(T t) {}
}
```

And the sub-class was:
```
class Subclass extends Parent<String> {}
```

Then the cloned method would be:
```
   public void inherited(String t) {}
```

If the super class has no type variables then the `createMethod()` will
call `PsiMethodItem.create(..., template)` to create a copy. Otherwise,
it will call `PsiMethodItem.toStubForCloning()` to resolve the type
variables if possible and then call
`PsiMethodItem.create(..., resolvedMethod)` to create a copy.

In the first case the copy will have exactly the same modifiers as the
original. In the second case the copy will always be marked `public`.
There seems to be no reason why the presence of type variables should
cause the modifier to be changed.

This change stops forcing the method to be public and fixes up the
couple of tests that were broken by the change but which do not seem to
be testing that specific functionality.

Bug: 292048098
Test: ./gradlew
Change-Id: I34539b487212b59c6d175c284bb5b9de40f9d91e