Add AnnotationInfo to cache annotation information
Previously, every time that some code needed to check a characteristic
of an annotation, e.g. whether it is a nullability annotation, it would
need to repeat the check itself. Wasting time and often complicating
the code by requiring code to directly access `options` properties.

This change adds a mechanism that allows for that information to be
efficiently cached and reused, saving time and simplifying the code.
It is currently only used  to cache the nullability status of the
annotation but follow up changes will use and extend that mechanism to
cache other information.

The information is cached in a `AnnotationInfo` which is added as a
lazily initialized property in the `DefaultAnnotationItem`. An instance
of that is provided by the `AnnotationManager` implementation when
needed.

The `BaseAnnotationManager` adds support for computing and caching the
`AnnotationItem` instances. It works by creating a key for an
`AnnotationItem` that differentiates it from all other
`AnnotationItem`s that may have a different `AnnotationItem`. At the
moment the key is always the `qualifiedName` as only that affects the
state of an `AnnotationItem`. Future changes will provide keys that
take into account attributes too as needed.

Bug: 292484829
Test: ./gradlew
Change-Id: Ia0be6f5c6514d387b66451b02d051b93b1336123